move persistent ui into its own folder
