+++ src/hooks/use-stores.test.ts
@@ -17,7 +17,7 @@ import {
 } from "./use-stores";
 import { unitConfigDefaults } from "../test-fixtures/sample-unit-configurations";
 import { UIModel } from "../models/stores/ui";
-import { PersistentUIModel } from "../models/stores/persistent-ui";
+import { PersistentUIModel } from "../models/stores/persistent-ui/persistent-ui";
 
 jest.mock("@concord-consortium/slate-editor", () => ({}));
